yamwi.php: base64 encode the "max" url parameter
Rationale: it is really nice to be able to exchange snippets via a
simple url. This change enables that.

The change does more: it changes the form "maxform" to be a GET method
form, by co-opting the submit button. This button no longer submits
the POST form data, it uses JS to replace the "max" URL parameter and
re-load the page with the new parameter.

TODO: avoid the race condition, where the same URL is simultaneously
requested by two or more agents.