Refactor button code
The code in comNN.js is repetitive, so we have re-written it by making
an array, buttons_action, whose entries are the relevant code
objects. A small function, make_buttons_action, creates a closure for
each code object. An event listener is added that maps
`make_buttons_action' over the array once the DOM is loaded and adds
the closure to the 'click' event listener of the corresponding button.