yamwi.mac: remove translate_into_{tex,print}, add yamwi_display
yamwi_display prints the 5 output types, each wrapped in the
appropriate html.

See also: show_output in yamwi.js.

There is a flaw in this set-up: the first opening elements must be
emitted by code passed into the script (see yamwi.php). The code does
work, but is brittle because of this.